requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
# ML dependencies
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
pymorphy3>=2.0.0
# Auth dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

